Added Controller for User routes and made some documentation
