+++ server/Controllers/Controller.md
@@ -1,0 +1,3 @@
+I think controller is supposed to make the code more modular so that the route.js files are cleaner to look at.
+
+This folder will have files that contains functions specific to their routes
